requirements.txt
python-telegram-bot==21.7
apscheduler==3.10.4
yt-dlp==2024.11.18
python-dotenv==1.0.1
pytz==2024.2
google-generativeai==0.8.5
